Allow simultaneous sensor history requests
This could occur if the user set up a list of futures with
sensor_history() requests and then yielded on that list.  There
is now unique state per request to keep track of these.

Unsubscribe from namespace once the download is completed, as it
won't be needed again.

See: CB-1584