Make _get_sitemap work with Tornado 6
HTTPClient is implemented with IOLoop.run_sync, and that can't be used
while another event loop is running on the same thread.

Convert _get_sitemap to a coroutine, and have the sitemap property run
it from a separate thread if necessary. This requires
concurrent.futures, so might not work on Python 2. It does not pass unit
tests because there are some mocks that get broken by this.