Re-jig way sitemap is initialised
My previous approach of adding `_init_sitemap` was error-prone because
one had to remember to add it everywhere the sitemap was used. I've
replaced it with a public `get_sitemap` coroutine that is used wherever
the sitemap property was previously used - so now one need only avoid
the sitemap property. In a few places I tweaked the code to call the
function only once and save the result instead of multiple times.

`sub_nr` also got similar treatment (since it has the same issues), with
a new `get_sub_nr` coroutine.